Update tasks.json
Added a new task called "test" that allows you to run tests for the web API project. This task runs the dotnet test command and uses the same project file as the other tasks.

Grouped the "build" task as the default build task. This means that when you run the build task, it will be the default task to execute if no other task is specified.